C interface + doc
git-svn-id: svn+ssh://nina.grame.fr/home/svn/guidoar@144 2a570e91-86af-4e14-8b94-b0b5b3996bc3

+++ src/visitors/event2timevisitor.cpp
@@ -76,7 +76,7 @@ void event2timevisitor::reset ()
 bool event2timevisitor::done ()
 {
 	fDone = false;
-	if (fTargetEvent >= 0) fDone = (fCountVisitor.currentCount() == fTargetEvent);
+	if (fTargetEvent >= 0) fDone = (fCountVisitor.currentCount() == (int)fTargetEvent);
 	else fDone = (currentVoiceDate() >= fTargetDate);
 	return fDone;
 }
